pyproject.toml
[project]
name = "n-bodies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "pygame>=2.6.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
simulation = "main:main"
